Support negative strides for zero-copy flip
Introduce signed Strides (int64_t) and add negative-stride support throughout the codebase to enable zero-copy flip views. Byte offsets are now computed with signed arithmetic, data() adjusts the base pointer for flipped views, required storage uses abs(stride), and contiguity checks account for negative strides. Implement flip as a zero-copy view by negating strides and add Tensor::has_negative_stride; update CPU operations and matmul/where/broadcast paths to use signed offsets. Add tests covering flip behavior, arithmetic on flipped tensors, and materialization via ascontiguousarray.

+++ include/axiom/shape.hpp
@@ -5,7 +5,7 @@
 namespace axiom {
 
 using Shape = std::vector<size_t>;
-using Strides = std::vector<size_t>;
+using Strides = std::vector<int64_t>;
 
 enum class MemoryOrder {
     RowMajor, // Row-major (C-style) - default
